Fix argument types

+++ tests/visualization_tests/test_edf.py
@@ -125,7 +125,7 @@ def test_empty_edf_info() -> None:
 
 
 @pytest.mark.parametrize("value", [float("inf"), -float("inf"), float("nan")])
-def test_nonfinite_removed(value: int) -> None:
+def test_nonfinite_removed(value: float) -> None:
 
     study = prepare_study_with_trials(value_for_first_trial=value)
     edf_info = _get_edf_info(study)
@@ -145,7 +145,7 @@ def test_nonfinite_removed(value: int) -> None:
 
 @pytest.mark.parametrize("objective", (0, 1))
 @pytest.mark.parametrize("value", (float("inf"), -float("inf")))
-def test_nonfinite_multiobjective(objective: int, value: int) -> None:
+def test_nonfinite_multiobjective(objective: int, value: float) -> None:
 
     study = prepare_study_with_trials(n_objectives=2, value_for_first_trial=value)
     edf_info = _get_edf_info(study, target=lambda t: t.values[objective])
